Update bibleDB Interface.py
Updated the multi-line input box for notes, so now it has a scrollbar, it can resize, and you can insert newlines anywhere in the text without accidentally appending one to the end of the text. Also now  the treeview will show chapters that have notes or tags in another color.