Update bibledb_Manager.py
Fixed some bugs related to yesterday's changes. Tidied up some verse navigation. Also made it so that if you click a tag in the tag-entry suggestion list, it will populate the entry-field with that tag, so if you click "OK" instead of double-clicking the tag in the entry list, it will go ahead and enter the tag you selected.